Move market configuration to separate file - Also test separately
